refactoring: APPLYALTER_PKG table handling extracted
