+++ src/main/java/ch/ips/g2/applyalter/ApplyAlter.java
@@ -135,6 +135,9 @@ public class ApplyAlter {
      */
     protected RunContext runContext;
 
+    @Nullable
+    private PkgLogTableHandler pkgLogTable;
+
     protected XStream xstream = new XStream();
     protected String username;
     protected Multimap<String, String> unapplied = ArrayListMultimap.create();
@@ -199,6 +202,10 @@ public class ApplyAlter {
         }
 
         determineEnvironment(runContext);
+
+        if (isLogTableUsed()) {
+            pkgLogTable = new PkgLogTableHandler(this.runContext, db, this.username);
+        }
     }
 
     private DbConfigFile parseConfiguration(String config) {
@@ -405,7 +412,9 @@ public class ApplyAlter {
 
             if (sourceHash != null) {
                 runContext.report(ALTER, "alterscripts done, source hash: %s", sourceHash);
-                savelog_pkg(sourceHash);
+                if (pkgLogTable != null) {
+                    pkgLogTable.savelog_pkg(sourceHash);
+                }
             }
         } finally {
             db.closeConnections();
@@ -692,47 +701,6 @@ public class ApplyAlter {
             s.executeUpdate();
         } catch (SQLException e) {
             runContext.report(ReportLevel.ERROR, "failed to insert applyalter_log record: %s", e.getMessage());
-        } finally {
-            DbUtils.close(s);
-        }
-    }
-
-    /**
-     * Log record to applyalter_pkg. Unlike record to applyalter_log, this one does commit!
-     *
-     * @param sourceHash hash of source bundle
-     */
-    protected void savelog_pkg(String sourceHash) {
-        if (sourceHash == null || runContext.getRunMode() != RunMode.SHARP || !isLogTableUsed()) {
-            //do not write to database
-            return;
-        }
-
-        for (DbInstance d : db.getEntries()) {
-            savelog_pkg(d, sourceHash);
-        }
-    }
-    /**
-     * Log record to applyalter_pkg. Unlike record to applyalter_log, this one does commit!
-     *
-     * @param d          database instance
-     * @param sourceHash hash of source bundle
-     */
-    private void savelog_pkg(DbInstance d, String sourceHash) {
-        String id = d.getId();
-        runContext.report(MAIN, "Package record: %s/%s", id, sourceHash);
-
-        Connection c = d.getConnection(runContext);
-        PreparedStatement s = null;
-        try {
-            s = c.prepareStatement("insert into " + d.getPkgLogTable() + " (dbid,hash,username) values (?,?,?)");
-            s.setString(1, id);
-            s.setString(2, sourceHash);
-            s.setString(3, username);
-            s.executeUpdate();
-            c.commit();
-        } catch (SQLException e) {
-            runContext.report(ReportLevel.ERROR, "failed to insert applyalter_pkg record: %s", e.getMessage());
         } finally {
             DbUtils.close(s);
         }
